refactor: move JiraSearchRequest model to requestModels
